\docType{package}
\name{dendextendRcpp-package}
\alias{Rcpp_count_leaves}
\alias{assign_dendextendRcpp_to_dendextend}
\alias{dendextendRcpp}
\alias{dendextendRcpp-package}
\alias{is_leaf}
\alias{is_list}
\alias{remove_dendextendRcpp_options}
\title{Functions for extending dendextend using Rcpp}
\description{
Functions for extending dendextend using Rcpp
}
\keyword{Rcpp}
\keyword{dendextend,}
\keyword{dendrogram,}
\keyword{package,}

